+++ payments-core/src/main/java/com/stripe/android/model/StripeIntent.kt
@@ -74,6 +74,9 @@ sealed interface StripeIntent : StripeModel {
 
     fun requiresConfirmation(): Boolean
 
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    fun getPaymentMethodOptions(): Map<String, Any?>
+
     /**
      * Type of the next action to perform.
      */