+++ payments-core/src/test/java/com/stripe/android/model/PaymentIntentTest.kt
@@ -274,4 +274,44 @@ class PaymentIntentTest {
         val result = paymentIntent.isSetupFutureUsageSet("card")
         assertThat(result).isFalse()
     }
+
+    @Test
+    fun `getPaymentMethodOptions returns expected results`() {
+        val paymentIntent = PaymentIntentFixtures.PI_REQUIRES_PAYMENT_METHOD.copy(
+            paymentMethodOptionsJsonString = """
+                {
+                    "card": {
+                        "mandate_options": null,
+                        "network": null,
+                        "request_three_d_secure": "automatic"
+                    },
+                    "us_bank_account": {
+                        "financial_connections": {
+                            "permissions": "balances"
+                        },
+                        "setup_future_usage": "on_session",
+                        "verification_method": "automatic"
+                    }
+                }
+            """.trimIndent()
+        )
+
+        assertThat(paymentIntent.getPaymentMethodOptions())
+            .isEqualTo(
+                mapOf(
+                    "card" to mapOf(
+                        "mandate_options" to null,
+                        "network" to null,
+                        "request_three_d_secure" to "automatic"
+                    ),
+                    "us_bank_account" to mapOf(
+                        "financial_connections" to mapOf(
+                            "permissions" to "balances"
+                        ),
+                        "setup_future_usage" to "on_session",
+                        "verification_method" to "automatic"
+                    )
+                )
+            )
+    }
 }