+++ payments-ui-core/src/test/java/com/stripe/android/ui/core/forms/resources/LpmRepositoryTest.kt
@@ -5,6 +5,7 @@ import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.model.PaymentMethod.Type.Card
 import com.stripe.android.model.PaymentMethod.Type.CashAppPay
+import com.stripe.android.model.PaymentMethod.Type.USBankAccount
 import com.stripe.android.paymentsheet.forms.Delayed
 import com.stripe.android.testing.PaymentIntentFactory
 import com.stripe.android.ui.core.CardBillingDetailsCollectionConfiguration
@@ -275,7 +276,15 @@ class LpmRepositoryTest {
     @Test
     fun `Verify that us_bank_account is supported when financial connections sdk available`() {
         lpmRepository.update(
-            PaymentIntentFactory.create(paymentMethodTypes = emptyList()),
+            PaymentIntentFactory.create(paymentMethodTypes = emptyList()).copy(
+                paymentMethodOptionsJsonString = """
+                    {
+                        "us_bank_account": {
+                            "verification_method": "automatic"
+                        }
+                    }
+                """.trimIndent()
+            ),
             """
               [
                 {
@@ -353,6 +362,69 @@ class LpmRepositoryTest {
 
         val supportedPaymentMethods = lpmRepository.values().map { it.code }
         assertThat(supportedPaymentMethods).containsExactly(Card.code, CashAppPay.code)
+    }
+
+    @Test
+    fun `Verify LpmRepository filters out USBankAccount if verification method is unsupported`() = runTest {
+        val lpmRepository = LpmRepository(
+            lpmInitialFormData = LpmRepository.LpmInitialFormData(),
+            arguments = LpmRepository.LpmRepositoryArguments(
+                resources = ApplicationProvider.getApplicationContext<Application>().resources,
+                enableACHV2InDeferredFlow = true
+            ),
+        )
+
+        val deferredPaymentIntent = PaymentIntentFactory.create(
+            paymentMethodTypes = listOf("card", "us_bank_account", "cashapp"),
+        ).copy(
+            clientSecret = null,
+            paymentMethodOptionsJsonString = """
+                {
+                    "us_bank_account": {
+                        "verification_method": "microdeposit"
+                    }
+                }
+            """.trimIndent()
+        )
+
+        lpmRepository.update(
+            stripeIntent = deferredPaymentIntent,
+            serverLpmSpecs = null,
+        )
+
+        val supportedPaymentMethods = lpmRepository.values().map { it.code }
+        assertThat(supportedPaymentMethods).containsExactly(Card.code, CashAppPay.code)
+    }
+
+    @Test
+    fun `Verify LpmRepository filters out USBankAccount if verification method is supported`() = runTest {
+        val lpmRepository = LpmRepository(
+            lpmInitialFormData = LpmRepository.LpmInitialFormData(),
+            arguments = LpmRepository.LpmRepositoryArguments(
+                resources = ApplicationProvider.getApplicationContext<Application>().resources,
+                enableACHV2InDeferredFlow = false
+            ),
+        )
+
+        val deferredPaymentIntent = PaymentIntentFactory.create(
+            paymentMethodTypes = listOf("card", "us_bank_account", "cashapp"),
+        ).copy(
+            paymentMethodOptionsJsonString = """
+                {
+                    "us_bank_account": {
+                        "verification_method": "automatic"
+                    }
+                }
+            """.trimIndent()
+        )
+
+        lpmRepository.update(
+            stripeIntent = deferredPaymentIntent,
+            serverLpmSpecs = null,
+        )
+
+        val supportedPaymentMethods = lpmRepository.values().map { it.code }
+        assertThat(supportedPaymentMethods).containsExactly(Card.code, USBankAccount.code, CashAppPay.code)
     }
 
     @Test